requirements.txt
django~=4.1.2
channels~=3.0.4
channels-redis~=3.2.0
asgiref~=3.5.2
python-telegram-bot~=13.7.0
webuntis~=0.1.21
Pillow~=9.2.0
firebase_admin~=5.0.2
py-firebase-dynamic-links~=1.0.1
django-cors-headers~=3.10.0